renamed folders that had similar project names. did not introduce any errors.
