+++ app/src/androidTest/java/com/example/favoritesdk/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.favoritesdkproject
+package com.example.favoritesdk
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
