+++ app/src/main/java/com/example/favoritesdk/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.favoritesdkproject.ui.theme
+package com.example.favoritesdk.ui.theme
 
 import androidx.compose.ui.graphics.Color
